Tidy up/remove redundant code
