+++ common/src/main/scala/models/GuardianItemType.scala
@@ -1,14 +1,17 @@
 package models
 
+import play.api.data.validation.ValidationError
 import play.api.libs.json.Json
 
 sealed case class GuardianItemType(mobileAggregatorPrefix: String)
+
 object GuardianItemType {
-  implicit val reads = Json.reads[GuardianItemType].map({
+  implicit val reads = Json.reads[GuardianItemType].collect(ValidationError("Unrecognised item type")) {
     case GuardianItemType("section") => GITSection
     case GuardianItemType("latest") => GITTag
     case GuardianItemType("item-trimmed") => GITContent
-  })
+  }
+
   implicit val writes = Json.writes[GuardianItemType]
 }
